RF: move all github helper functions into datalad/support/github_.py
Relevant tests also were moved accordingly
